Add waveform PTP metric
